Added generator.py and caller.py
The Generator file is basically used to portray how the voice is generated. The main part is in Caller file, where the user can select the caller they want to speak to.
Note: Remember to replace the API key and the Gemini Model's name to make the model work correctly.